requirements.txt
flask
streamlit
google-generativeai
anthropic
groq
openai
faiss-cpu
numpy
pandas
requests
ragas
datasets
umap-learn
plotly
scikit-learn
